+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## WIP
+## 1.3.0
 - Add French translations (thanks @MATsxm)
 - Fix loading Google Fonts on error and offline pages
 - Fix initial value for hide-component parameter